43. Multiply Strings
